move to usb_comms.h
